+++ Project/src/main/java/com/lukecreator/BonziBot/Commands/Admin/SetRepCommand.java
@@ -28,7 +28,7 @@ public class SetRepCommand extends Command {
 	@Override
 	public void run(CommandExecutionInfo e) {
 		User target = e.args.getUser("target");
-		int newRep = e.args.getInt("new rep");
+		int newRep = (int) e.args.getInt("new rep");
 		
 		UserAccountManager uam = e.bonzi.accounts;
 		UserAccount account = uam.getUserAccount(target);
